+++ client/src/components/RatingsAndReviews/Ratings.jsx
@@ -1,3 +1,34 @@
 import React from 'react';
 import axios from 'axios';
 import styled from 'styled-components';
+
+const Ratings = () => (
+  <div>
+    <span>
+      100% of reviews recommend this product
+      <br />
+    </span>
+    <span>
+      5 stars
+      <br />
+    </span>
+    <span>
+      4 stars
+      <br />
+    </span>
+    <span>
+      3 stars
+      <br />
+    </span>
+    <span>
+      2 stars
+      <br />
+    </span>
+    <span>
+      1 star
+      <br />
+    </span>
+  </div>
+);
+
+export default Ratings;
